refactor: internaliza configuracao do logger
